requirements.txt
Django>=6.0
djangorestframework>=3.15
django-cors-headers>=4.0
django-ckeditor>=6.7
django-filter>=24.0
Pillow>=10.0
psycopg2-binary>=2.9
requests>=2.31
